Fix #6592: Answer shields p3a question with initial state once
This matches desktops behaviour where they seem to record the answer only once
