+++ Client/Frontend/Shields/ShieldsViewController.swift
@@ -27,7 +27,7 @@ class ShieldsViewController: UIViewController, PopoverContentComponent {
     return _url
   }()
 
-  var shieldsSettingsChanged: ((ShieldsViewController) -> Void)?
+  var shieldsSettingsChanged: ((ShieldsViewController, BraveShield) -> Void)?
   var showGlobalShieldsSettings: ((ShieldsViewController) -> Void)?
 
   private var statsUpdateObservable: AnyObject?
@@ -105,10 +105,6 @@ class ShieldsViewController: UIViewController, PopoverContentComponent {
     Domain.setBraveShield(
       forUrl: url, shield: shield, isOn: isOn,
       isPrivateBrowsing: PrivateBrowsingManager.shared.isPrivateBrowsing)
-    DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
-      // Record shields & FP related hisotgrams, wait a sec for CoreData to sync contexts
-      self.recordShieldsUpdateP3A(shield: shield)
-    }
   }
 
   private func updateGlobalShieldState(_ on: Bool, animated: Bool = false) {
@@ -273,7 +269,7 @@ class ShieldsViewController: UIViewController, PopoverContentComponent {
         // Wait a fraction of a second to allow DB write to complete otherwise it will not use the
         // updated shield settings when reloading the page
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-          self.shieldsSettingsChanged?(self)
+          self.shieldsSettingsChanged?(self, shield)
         }
       }
     }
@@ -286,7 +282,7 @@ class ShieldsViewController: UIViewController, PopoverContentComponent {
     // Wait a fraction of a second to allow DB write to complete otherwise it will not use the updated
     // shield settings when reloading the page
     DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-      self.shieldsSettingsChanged?(self)
+      self.shieldsSettingsChanged?(self, .AllOff)
     }
   }
 
@@ -343,41 +339,5 @@ class ShieldsViewController: UIViewController, PopoverContentComponent {
   @available(*, unavailable)
   required init?(coder aDecoder: NSCoder) {
     fatalError()
-  }
-  
-  // MARK: - P3A
-  
-  private enum P3AShieldUpdateKind {
-    case adblock
-    case fingerprintingProtection
-  }
-  
-  private func recordShieldsUpdateP3A(shield: BraveShield) {
-    let buckets: [Bucket] = [
-      0,
-      .r(1...5),
-      .r(6...10),
-      .r(11...20),
-      .r(21...30),
-      .r(31...),
-    ]
-    switch shield {
-    case .AdblockAndTp:
-      // Q51 On how many domains has the user set the adblock setting to be lower (block less) than the default?
-      let adsBelowGlobalCount = Domain.totalDomainsWithAdblockShieldsLoweredFromGlobal()
-      UmaHistogramRecordValueToBucket("Brave.Shields.DomainAdsSettingsBelowGlobal", buckets: buckets, value: adsBelowGlobalCount)
-      // Q52 On how many domains has the user set the adblock setting to be higher (block more) than the default?
-      let adsAboveGlobalCount = Domain.totalDomainsWithAdblockShieldsIncreasedFromGlobal()
-      UmaHistogramRecordValueToBucket("Brave.Shields.DomainAdsSettingsAboveGlobal", buckets: buckets, value: adsBelowGlobalCount)
-    case .FpProtection:
-      // Q53 On how many domains has the user set the FP setting to be lower (block less) than the default?
-      let fingerprintingBelowGlobalCount = Domain.totalDomainsWithFingerprintingProtectionLoweredFromGlobal()
-      UmaHistogramRecordValueToBucket("Brave.Shields.DomainFingerprintSettingsBelowGlobal", buckets: buckets, value: fingerprintingBelowGlobalCount)
-      // Q54 On how many domains has the user set the FP setting to be higher (block more) than the default?
-      let fingerprintingAboveGlobalCount = Domain.totalDomainsWithFingerprintingProtectionIncreasedFromGlobal()
-      UmaHistogramRecordValueToBucket("Brave.Shields.DomainFingerprintSettingsAboveGlobal", buckets: buckets, value: fingerprintingAboveGlobalCount)
-    case .AllOff, .NoScript, .SafeBrowsing:
-      break
-    }
   }
 }
